+++ Data/models/Questions/QuestionsChoices.cs
@@ -22,7 +22,7 @@ namespace Data.models.Questions
 
         public bool IsRightAnswer { get; set; }
 
-        public decimal Rank { get; set; }
+        public byte Rank { get; set; }
 
 
     }
